requirements.txt
Authlib~=1.0.1
certifi~=2020.6.20
cffi~=1.15.0
chardet~=3.0.4
click~=7.1.2
cryptography~=36.0.2
dominate~=2.5.2
Flask~=1.1.2
Flask-Bootstrap~=3.3.7.1
Flask-SQLAlchemy~=2.5.1
Flask-WTF~=0.14.3
greenlet~=1.1.0
gunicorn~=20.1.0
idna~=2.10
itsdangerous~=1.1.0
Jinja2~=2.11.2
MarkupSafe~=1.1.1
numpy~=1.21.1
Pillow~=8.3.2
pip~=20.2.3
psycopg2-binary~=2.9.1
pycparser~=2.21
python-dotenv~=0.19.0
requests~=2.24.0
SQLAlchemy~=1.4.22
urllib3~=1.25.10
visitor~=0.1.3
